#28 - Move the request, response file path into the dto
request, response는 dto에 포함되므로 안으로 경로를 이동
