+++ src/main/java/com/board/projectboard/controller/ArticleController.java
@@ -3,9 +3,9 @@ package com.board.projectboard.controller;
 import com.board.projectboard.domain.constant.FormStatus;
 import com.board.projectboard.domain.constant.SearchType;
 import com.board.projectboard.dto.UserAccountDto;
-import com.board.projectboard.request.ArticleRequest;
-import com.board.projectboard.response.ArticleResponse;
-import com.board.projectboard.response.ArticleWithCommentsResponse;
+import com.board.projectboard.dto.request.ArticleRequest;
+import com.board.projectboard.dto.response.ArticleResponse;
+import com.board.projectboard.dto.response.ArticleWithCommentsResponse;
 import com.board.projectboard.service.ArticleService;
 import com.board.projectboard.service.PaginationService;
 import lombok.RequiredArgsConstructor;
