+++ src/main/java/com/board/projectboard/dto/response/ArticleResponse.java
@@ -1,8 +1,7 @@
-package com.board.projectboard.response;
+package com.board.projectboard.dto.response;
 
 import com.board.projectboard.dto.ArticleDto;
 
-import java.io.Serializable;
 import java.time.LocalDateTime;
 
 public record ArticleResponse(
