+++ src/test/java/com/board/projectboard/controller/ArticleControllerTest.java
@@ -6,12 +6,11 @@ import com.board.projectboard.domain.constant.SearchType;
 import com.board.projectboard.dto.ArticleDto;
 import com.board.projectboard.dto.ArticleWithCommentsDto;
 import com.board.projectboard.dto.UserAccountDto;
-import com.board.projectboard.request.ArticleRequest;
-import com.board.projectboard.response.ArticleResponse;
+import com.board.projectboard.dto.request.ArticleRequest;
+import com.board.projectboard.dto.response.ArticleResponse;
 import com.board.projectboard.service.ArticleService;
 import com.board.projectboard.service.PaginationService;
 import com.board.projectboard.util.FormDataEncoder;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
